+++ projects/wbt101key/ch04a/AdvScanner/AdvScanner.c
@@ -1,58 +1,57 @@
-/*
- * This file has been automatically generated by the WICED 20719-B1 Designer.
- * Bluetooth Application.
- *
- */
-
-/** AdvScanner.c
- *
- */
-
-#include "wiced.h"
-#include "wiced_bt_dev.h"
-#include "wiced_bt_ble.h"
-#include "wiced_bt_uuid.h"
-#include "wiced_bt_gatt.h"
-#include "wiced_hal_nvram.h"
-#include "wiced_hal_gpio.h"
-#include "wiced_bt_app_hal_common.h"
-#include "wiced_hal_platform.h"
-#include "wiced_bt_trace.h"
+/*******************************************************************
+ * Imports
+ ******************************************************************/
 #include "sparcommon.h"
 #include "hci_control_api.h"
-#include "wiced_transport.h"
+
+#include "wiced.h"
+#include "wiced_bt_app_hal_common.h"
+#include "wiced_bt_ble.h"
+#include "wiced_bt_cfg.h"
+#include "wiced_bt_dev.h"
+#include "wiced_bt_gatt.h"
+#include "wiced_bt_stack.h"
+#include "wiced_bt_trace.h"
+#include "wiced_bt_uuid.h"
+#include "wiced_hal_gpio.h"
+#include "wiced_hal_nvram.h"
+#include "wiced_hal_platform.h"
 #include "wiced_hal_pspi.h"
 #include "wiced_hal_puart.h"
-#include "wiced_bt_cfg.h"
-#include "company_ids.h"
 #include "wiced_rtos.h"
+#include "wiced_transport.h"
+
+#include "company_ids.h"
 #include "device_table.h"
-#include "wiced_bt_stack.h"
+#include "print_functions.h"
 
 /*******************************************************************
  * Constant Definitions
  ******************************************************************/
 #define TRANS_UART_BUFFER_SIZE  1024
 #define TRANS_UART_BUFFER_COUNT 2
+#define TYPE_TIMEOUT            2
 
 /*******************************************************************
  * Variable Definitions
  ******************************************************************/
 extern const wiced_bt_cfg_settings_t wiced_bt_cfg_settings;
 extern const wiced_bt_cfg_buf_pool_t wiced_bt_cfg_buf_pools[WICED_BT_CFG_NUM_BUF_POOLS];
-// Local Device Name in device configuration
-extern uint8_t BT_LOCAL_NAME[];
 // Transport pool for sending RFCOMM data to host
 static wiced_transport_buffer_pool_t* transport_pool = NULL;
 static wiced_bt_ble_scan_type_t scan_state = BTM_BLE_SCAN_TYPE_NONE;
+
 static wiced_bool_t print = WICED_TRUE;
+static wiced_bool_t old_print;
+static wiced_bool_t table_on = WICED_FALSE;
 static wiced_bool_t filter = WICED_FALSE;
 static uint32_t timer = 0;
+static uint32_t type_time = 0;
 
 /*******************************************************************
  * Function Prototypes
  ******************************************************************/
-static void                  advscanner_change_scan_state      (wiced_bt_ble_scan_type_t new_state);
+static void                  change_scan_state      (wiced_bt_ble_scan_type_t new_state);
 static wiced_bt_dev_status_t advscanner_management_callback    ( wiced_bt_management_evt_t event, wiced_bt_management_evt_data_t *p_event_data );
 static uint32_t              hci_control_process_rx_cmd        ( uint8_t* p_data, uint32_t len );
 static void                  rx_cback                          ( void *data );
@@ -146,10 +145,8 @@ void newAdv(wiced_bt_ble_scan_results_t *p_scan_result, uint8_t *p_adv_data)
     (void)dt_addDevice(p_scan_result,p_adv_data,timer);
 }
 
-/*
- * This function is executed in the BTM_ENABLED_EVT management callback.
- */
-static void advscanner_change_scan_state(wiced_bt_ble_scan_type_t new_state)
+/* This function is executed in the BTM_ENABLED_EVT management callback */
+static void change_scan_state(wiced_bt_ble_scan_type_t new_state)
 {
     if(new_state != scan_state)
     {
@@ -157,7 +154,6 @@ static void advscanner_change_scan_state(wiced_bt_ble_scan_type_t new_state)
         scan_state = new_state;
     }
 }
-
 
 /* Bluetooth Management Event Handler */
 wiced_bt_dev_status_t advscanner_management_callback( wiced_bt_management_evt_t event, wiced_bt_management_evt_data_t *p_event_data )
@@ -191,7 +187,7 @@ wiced_bt_dev_status_t advscanner_management_callback( wiced_bt_management_evt_t 
             /* Enables scanning */
             timer_thread = wiced_rtos_create_thread();
             wiced_rtos_init_thread(timer_thread,4,"timer",timer_function,1000,0);
-            advscanner_change_scan_state(BTM_BLE_SCAN_TYPE_HIGH_DUTY);
+            change_scan_state(BTM_BLE_SCAN_TYPE_HIGH_DUTY);
         }
         /* Initialize the UART for input */
         wiced_hal_puart_init( );
@@ -287,16 +283,6 @@ wiced_bt_dev_status_t advscanner_management_callback( wiced_bt_management_evt_t 
     return status;
 }
 
-void clear_terminal()
-{
-    uint8_t seq1[4] = {0x1B,'[','2','J'};
-    // Esc[2J clears the screen
-    wiced_hal_puart_print((char*)seq1);
-    // Esc[H moves the cursor to the top left corner
-    uint8_t seq2[3] = {0x1B,'[','H'};
-    wiced_hal_puart_print((char*)seq2);
-}
-
 /* Interrupt callback function for UART */
 void rx_cback( void *data )
 {
@@ -307,104 +293,146 @@ void rx_cback( void *data )
     wiced_hal_puart_read( &readbyte );
     wiced_hal_puart_reset_puart_interrupt();
 
-    if( readbyte == 's' )
+    if(!table_on)
     {
-        /* Turn scanning on */
-        advscanner_change_scan_state(BTM_BLE_SCAN_TYPE_HIGH_DUTY);
+        /* Process the read character */
+        switch (readbyte)
+        {
+        case 's':
+            /* Turn scanning on */
+            change_scan_state(BTM_BLE_SCAN_TYPE_HIGH_DUTY);
+            break;
+        case 'S':
+            /* Turn scanning off */
+            change_scan_state(BTM_BLE_SCAN_TYPE_NONE);
+            break;
+        case 't':
+            /* Dump table */
+            wiced_hal_puart_print( "\n\r" );
+            printDeviceTableOneLine();
+            wiced_hal_puart_print( "\n\r" );
+            break;
+        case 'm':
+            /* Dump multiline table */
+            old_print = print;
+            print = WICED_FALSE;
+            table_on = WICED_TRUE;
+            printDeviceTableMultiLine();
+            break;
+        case 'f':
+            /* Enabled filter */
+            if(filter == WICED_FALSE)
+                wiced_hal_puart_print( "Filter Enabled\n\r" );
+            filter = WICED_TRUE;
+            break;
+        case 'F':
+            /* Disable filter */
+            if(filter == WICED_TRUE)
+                wiced_hal_puart_print( "Filter Disabled\n\r" );
+            filter = WICED_FALSE;
+            break;
+        case 'p':
+            /* Enable printing */
+            if(print == WICED_FALSE)
+                wiced_hal_puart_print( "Printing Enabled\n\r" );
+            print = WICED_TRUE;
+            break;
+        case 'P':
+            /* Disable printing */
+            if(print == WICED_TRUE)
+                wiced_hal_puart_print( "Printing Disabled\n\r" );
+            print = WICED_FALSE;
+            break;
+        case 'd':
+            /* Clear device data */
+            timer = 0;
+            dt_reset();
+            wiced_hal_puart_print( "All device data cleared\n\r" );
+            clear_terminal();
+            break;
+        case 'c':
+            /* Clear the terminal */
+            clear_terminal();
+            break;
+        case 'r':
+            /* Dump recent filtered packets */
+            printRecentFilterData();
+            break;
+        case '?':
+            /* Print help */
+            wiced_hal_puart_print( "\n\r" );
+            wiced_hal_puart_print( "+-------- Available Commands --------+\n\r" );
+            wiced_hal_puart_print( "|  s/S    Enable/Disable Scanning    |\n\r" );
+            wiced_hal_puart_print( "|  p/P    Enable/Disable Printing    |\n\r" );
+            wiced_hal_puart_print( "|  f/F    Enable/Disable Filter      |\n\r" );
+            wiced_hal_puart_print( "|  0-9    Change Filtered Device     |\n\r" );
+            wiced_hal_puart_print( "|  r      Dump Recent Filter Packets |\n\r" );
+            wiced_hal_puart_print( "|  t      Dump Single-line Table     |\n\r" );
+            wiced_hal_puart_print( "|  m      Dump Multiline Table       |\n\r" );
+            wiced_hal_puart_print( "|  d      Delete All Device Data     |\n\r" );
+            wiced_hal_puart_print( "|  c      Clear Screen               |\n\r" );
+            wiced_hal_puart_print( "|  ?      Print Commands             |\n\r" );
+            wiced_hal_puart_print( "+------------------------------------+\n\r" );
+            wiced_hal_puart_print( "\n\r" );
+            break;
+        case '0':
+        case '1':
+        case '2':
+        case '3':
+        case '4':
+        case '5':
+        case '6':
+        case '7':
+        case '8':
+        case '9':
+            /* Set filter number */
+            if(timer - type_time > TYPE_TIMEOUT)
+                dt_setFocus(readbyte-'0');
+            else
+                dt_setFocus(10*dt_getFocus()+(readbyte-'0'));
+            type_time = timer;
+            break;
+        }
     }
-    else if( readbyte == 'S' )
+    else
     {
-        /* Turn scanning off */
-        advscanner_change_scan_state(BTM_BLE_SCAN_TYPE_NONE);
-    }
-    else if( readbyte == 't' )
-    {
-        /* Dump table */
-        wiced_hal_puart_print( "\n\r" );
-        dt_printDeviceTableOneLine( );
-        wiced_hal_puart_print( "\n\r" );
-    }
-    else if( readbyte == 'm' )
-    {
-        /* Dump multiline table */
-        wiced_hal_puart_print( "\n\r" );
-        dt_printDeviceTableMultiLine( );
-        wiced_hal_puart_print( "\n\r" );
-    }
-    else if( readbyte == 'f' )
-    {
-        /* Enabled filter */
-        if(filter == WICED_FALSE)
-            wiced_hal_puart_print( "Filter Enabled\n\r" );
-        filter = WICED_TRUE;
-    }
-    else if( readbyte == 'F' )
-    {
-        /* Disable filter */
-        if(filter == WICED_TRUE)
-            wiced_hal_puart_print( "Filter Disabled\n\r" );
-        filter = WICED_FALSE;
-    }
-    else if( readbyte == 'p' )
-    {
-        /* Enable printing */
-        if(print == WICED_FALSE)
-            wiced_hal_puart_print( "Printing Enabled\n\r" );
-        print = WICED_TRUE;
-    }
-    else if( readbyte == 'P' )
-    {
-        /* Disable printing */
-        if(print == WICED_TRUE)
-            wiced_hal_puart_print( "Printing Disabled\n\r" );
-        print = WICED_FALSE;
-    }
-    else if( readbyte == 'd' )
-    {
-        /* Clear device data */
-        timer = 0;
-        dt_reset();
-        wiced_hal_puart_print( "All device data cleared\n\r" );
-        clear_terminal();
-    }
-    else if( readbyte == 'c')
-    {
-        /* Clear the terminal */
-        clear_terminal();
-    }
-    else if( readbyte == 'r')
-    {
-        /* Dump recent filtered packets */
-        dt_printRecentFilterData();
-    }
-    else if( readbyte == '?' )
-    {
-        /* Print help */
-        wiced_hal_puart_print( "\n\r" );
-        wiced_hal_puart_print( "+-------- Available Commands --------+\n\r" );
-        wiced_hal_puart_print( "|  s/S    Enable/Disable Scanning    |\n\r" );
-        wiced_hal_puart_print( "|  p/P    Enable/Disable Printing    |\n\r" );
-        wiced_hal_puart_print( "|  f/F    Enable/Disable Filter      |\n\r" );
-        wiced_hal_puart_print( "|  r      Dump Recent Filter Packets |\n\r" );
-        wiced_hal_puart_print( "|  t      Dump Single-line Table     |\n\r" );
-        wiced_hal_puart_print( "|  m      Dump Multiline Table       |\n\r" );
-        wiced_hal_puart_print( "|  d      Delete All Device Data     |\n\r" );
-        wiced_hal_puart_print( "|  c      Clear Screen               |\n\r" );
-        wiced_hal_puart_print( "|  ?      Print Commands             |\n\r" );
-        wiced_hal_puart_print( "+------------------------------------+\n\r" );
-        wiced_hal_puart_print( "\n\r" );
-    }
-    else if( readbyte == '2')
-    {
-        focus = dt_getFocus();
-        dt_setFocus(focus+1);
-        WICED_BT_TRACE( "Filtering on device %d\n\r", dt_getFocus() );
-    }
-    else if( readbyte == '1' )
-    {
-        focus = dt_getFocus();
-        dt_setFocus(focus-1);
-        WICED_BT_TRACE( "Filtering on device %d\n\r", dt_getFocus() );
+        /* Process the read character */
+        switch (readbyte)
+        {
+        /* Change table pages */
+        case '>':
+            if(table_on == WICED_TRUE)
+            {
+                incrementPageNum();
+                printDeviceTableMultiLine();
+            }
+            break;
+        case '<':
+            if(table_on == WICED_TRUE)
+            {
+                decrementPageNum();
+                printDeviceTableMultiLine();
+            }
+            break;
+        case 0x1B:
+        case 'c':
+            /* Clear the terminal and exit the table */
+            table_on = WICED_FALSE;
+            print = old_print;
+            clear_terminal();
+            break;
+        case '?':
+            /* Print help */
+            wiced_hal_puart_print( "\n\r" );
+            wiced_hal_puart_print( "+-- Available Commands --+\n\r" );
+            wiced_hal_puart_print( "|  <     Increment Page  |\n\r" );
+            wiced_hal_puart_print( "|  >     Decrement Page  |\n\r" );
+            wiced_hal_puart_print( "|  ESC   Exit Table      |\n\r" );
+            wiced_hal_puart_print( "|  ?     Print Commands  |\n\r" );
+            wiced_hal_puart_print( "+------------------------+\n\r" );
+            wiced_hal_puart_print( "\n\r" );
+            break;
+        }
     }
 }
