+++ projects/wbt101key/ch04a/AdvScanner/AdvScanner.h
@@ -1,6 +1,4 @@
 #pragma once
-#include "wiced.h"
-#include "wiced_bt_ble.h"
 
 wiced_bool_t printing_enabled(void);
 wiced_bool_t filter_enabled(void);
