+++ projects/wbt101key/ch04a/AdvScanner/company_ids.c
@@ -1,6 +1,5 @@
 #include "wiced.h"
 #include "company_ids.h"
-#include "wiced_bt_trace.h"
 
 typedef struct {
 uint16_t companyCode;
@@ -1645,15 +1644,9 @@ uint32_t numCompanies = sizeof(company_ids)/sizeof(company_id_t);
 
 char *getCompanyName(uint16_t id)
 {
-    //WICED_BT_TRACE("Looking for = %X\n\r",id);
     for(int i=0;i<numCompanies;i++)
-    {
         if(company_ids[i].companyCode == id)
-        {
-            //WICED_BT_TRACE("Found %i id=%x name=%s\n\r",i,company_ids[i].companyCode,company_ids[i].name);
             return company_ids[i].name;
-        }
-    }
     return 0;
 }
