+++ projects/wbt101key/ch04a/AdvScanner/company_ids.h
@@ -1,4 +1,3 @@
 #pragma once
 
-
 char *getCompanyName(uint16_t id);
