+++ projects/wbt101key/ch04a/AdvScanner/decode_functions.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void bleDecodeAdInfo(uint8_t *bytes);
