+++ projects/wbt101key/ch04a/AdvScanner/ring_buffer.h
@@ -1,0 +1,3 @@
+void rb_insert(uint8_t *data);
+void rb_reset();
+void rb_print();
